Refactor data generation in callback function
